Fix frontapp aggregated row handling
The "aggregated" row of frontapp analytics API is not always at the beginning
of the returned rows.

In some cases, it is at the end of the array.

Hence, we change a little the logic to apply the "default" values when needed,
regardless of where we are in the returned array to be more robust.